Rename camera exported variable for clarity
this._params.camera in player.js needed to be kept the same.
It is an object that is always called camera as seen in the
debugger for the elements of the _params array. It should not
reference the .firstPersonCamera where .camera used to be like
the other change I made.