Remove floorContainer body
We do not need a body for the floorContainer. We need a single body for
the ground on which everything stands, which this floo container was
representing anyways. We do not need a body for multiple floorContainers either,
for which there are multiple. The floorContainers are simply for the
floors that stand on the ground. They might have physics in future.